requirements.txt
phidata
pyyaml
plotly
yfinance
streamlit
dataclasses
python-dotenv
google-generativeai